+++ examples/github-notification-filters/searchClient.ts
@@ -1,4 +1,4 @@
-import algoliasearch from 'algoliasearch';
+import algoliasearch from 'algoliasearch/lite';
 
 const appId = 'latency';
 const apiKey = '147a0e7dbc37d4c4dec9ec31b0f68716';