+++ examples/two-column-layout/src/plugins/recentSearchesPlugin.tsx
@@ -1,0 +1,11 @@
+import {
+  createLocalStorageRecentSearchesPlugin,
+  search,
+} from '@algolia/autocomplete-plugin-recent-searches';
+
+export const recentSearchesPlugin = createLocalStorageRecentSearchesPlugin({
+  key: 'autocomplete-two-column-layout-example',
+  search(params) {
+    return search({ ...params, limit: params.query ? 1 : 4 });
+  },
+});