+++ examples/two-column-layout/src/types/FaqHit.ts
@@ -1,0 +1,9 @@
+import { AutocompleteHit } from './AutocompleteHit';
+
+type FaqRecord = {
+  list_categories: string[];
+  title: string;
+  description: string;
+};
+
+export type FaqHit = AutocompleteHit<FaqRecord>;