+++ examples/two-column-layout/src/utils/isDetached.ts
@@ -1,0 +1,7 @@
+export function isDetached() {
+  return window.matchMedia(
+    getComputedStyle(document.documentElement).getPropertyValue(
+      '--aa-detached-media-query'
+    )
+  ).matches;
+}